LayoutTests/fast/text/international/system-language/navigator-language/navigator-language-es-MX-expected.txt

PASS navigator.language is "es-mx"
PASS successfullyParsed is true

TEST COMPLETE
